+++ include/plotcraft/Config.h
@@ -1,11 +1,11 @@
 #pragma once
 #include "Global.h"
-#include "Version.h"
 #include "ll/api/Config.h"
-#include "plotcraft/Macro.h"
+#include "plotcraft/Global.h"
 #include "plotcraft/utils/EconomySystem.h"
 #include <string>
 #include <vector>
+
 
 
 namespace plo {
