--- include/plotcraft/Macro.h
@@ -1,12 +1,0 @@
-#pragma once
-
-
-#ifdef PLOT_EXPORTS
-
-#define PLAPI __declspec(dllexport)
-
-#else
-
-#define PLAPI __declspec(dllimport)
-
-#endif
