--- include/plotcraft/Version.h
@@ -1,9 +1,0 @@
-
-// Config
-#define CONFIG_VERSION 7
-
-// PlotMetadata
-#define METADATA_VERSION 5
-
-// PlayerSetting
-#define SETTING_VERSION 1
