+++ include/plotcraft/data/PlayerNameDB.h
@@ -3,7 +3,7 @@
 #include "ll/api/data/KeyValueDB.h"
 #include "mc/deps/core/mce/UUID.h"
 #include "mc/world/actor/player/Player.h"
-#include "plotcraft/Macro.h"
+#include "plotcraft/Global.h"
 
 
 using string = std::string;
