+++ include/plotcraft/data/PlotDBStorage.h
@@ -1,9 +1,8 @@
 #pragma once
 #include "ll/api/data/KeyValueDB.h"
-#include "plotcraft/Macro.h"
-#include "plotcraft/Version.h"
-#include "plotcraft/core/PPos.h"
+#include "plotcraft/Global.h"
 #include "plotcraft/data/PlotMetadata.h"
+#include "plotcraft/math/PPos.h"
 #include <cstddef>
 #include <memory>
 #include <shared_mutex>
