+++ include/plotcraft/utils/EconomySystem.h
@@ -2,7 +2,7 @@
 #include "Date.h"
 #include "Utils.h"
 #include "mc/world/actor/player/Player.h"
-#include "plotcraft/Macro.h"
+#include "plotcraft/Global.h"
 
 using string = std::string;
