+++ src/plotcraft/command/Command.cc
@@ -21,6 +21,7 @@
 #include <unordered_map>
 
 
+
 namespace plo::command {
 using namespace plo::utils;
 using namespace plo::mc;
